Refactor Signature, split in Stack. Add sleep range in bucket.
Split Signature in two, Signature and Stack. This makes the separation more
logical. Add SleepMin and SleepMax, so that a range of sleep can be properly
presented to the user. Update main to present this range.

This is a big refactor so there's a risk I messed up. Please contact me if you
see anything suspicious.

Fixes #9