basic filter implementation
this addresses a common need for me. I often want to restrict output to
a certain package or set of files because i know that's where the stuff
is happening that I'm looking for.

here's some ideas to make this feature more advanced:
* both conditions are currently OR'ed. may want to change to AND, or
  leave it up to user. we may also want to support multiple conditions.
  something like the below perhaps:
  `-filter "path=foo && (package == bar || package == baz)"`
* instead of substring/equality, may want to support regex, especially for path

... I haven't personally had this need though, so I would keep it simple
until it's clear we want to support such fancier things.