[stack] Add Call.ImportPath() and Call.RelSrcPath
They are only set with pp -rebase / ParseDump(..., guesspath=true).

Expose better relative paths. That is, relative to either GOROOT or
GOPATH. This can be keyed on IsStdlib.

- Add -rel-path to add relative paths to GOROOT/GOPATH as a middle
  ground between the default behavior and -full-path.
- Add logic to support this in stack.

I hope to eventually make it so that it can be "guessed" without file
system access. This is necessary for good performance in webstack.