panicweb: add a go module in the stack trace
Use a callback in colorable to be able to include a go module file path.
It can't be in panic because importing colorable causes unexpected
goroutine assembly routine, similar to when net is imported.

This permits to start experimenting with properly parsing go module
paths (and the version).

Improve the expectations.