stack: call Augment() in TestAugment
Duh. That said it doesn't work because of a bug in the parsing code when
running a program via go run where the source is outside GOPATH and
doesn't have a go.mod.

Use a defer for cleaning since it's now using a t.Run(), so the code is
scoped. Can't use testing.T.Cleanup() since we run tests on go1.9.