panic: reduce static call depth from 2000 to 497
Anything longer fails on go1.18. As part of the research, I discovered
that if the functions are put in a different file, which I did to do a
different compilation between go1.17 and go1.18, it failed to link all
the way back to go1.9.7, which is the earliest version that is still
tested on.

Fixes #75