Avoid unicode filename
This is similar to 53e9d3ee100e0cf54d18eeae09dda6e09a29d090.

Over at https://github.com/cockroachdb/cockroach, we use bazel and it
chokes on this file:

```
java.io.IOException: Error extracting
[...]/com_github_maruel_panicparse_v2-v2.2.1.zip to [...]:
[...]/com_github_maruel_panicparse_v2/cmd/panic/internal/utf8/?tf8.go
(Illegal byte sequence)
```

which is why we're on a fork so far. It would be nice to avoid this.

There is an (eternal, it seems) [upstream discussion], so there is
little home of bazel starting to support non-unicode filenames.

I wonder if the loss in test coverage (if any) is acceptable in return
for avoiding build problems like the ones we've been seeing.

[upstream discussion]: https://github.com/bazelbuild/bazel/issues/374