stack: Fix test expectations for go1.19 and update GA to go1.19
Bump github actions to run on go1.19.

When starting a program on macOS on go1.18 and earlier, it seems that
exec.Cmd.Run() used to evaluate symlinks. It is not true anymore on
go1.19. Update test expectations accordingly.

go1.19 also changed stack traces to not include assembly like it did
previously.

Test only change.